CountryServiceImplTest: adapted tests after changing countByName() return type.
Correction for f2d23795d86caee40e0ecf2488d6115253f31fb1 commit.

No functional changes.

+++ src/test/groovy/ru/mystamps/web/service/CountryServiceImplTest.groovy
@@ -282,11 +282,11 @@ class CountryServiceImplTest extends Specification {
 	
 	def "countByName() should call dao"() {
 		given:
-			jdbcCountryDao.countByName(_ as String) >> 2
+			jdbcCountryDao.countByName(_ as String) >> 2L
 		when:
-			int result = service.countByName('Any name here')
+			long result = service.countByName('Any name here')
 		then:
-			result == 2
+			result == 2L
 	}
 	
 	def "countByName() should pass country name to dao"() {
